Renamed files and moved them around
