+++ Software/real_time/ROS_RoboBuggy/src/robobuggy/CMakeLists.txt
@@ -144,14 +144,14 @@ include_directories(
 ## With catkin_make all packages are built within a single CMake context
 ## The recommended prefix ensures that target names across packages don't collide
 # add_executable(${PROJECT_NAME}_node src/robobuggy_node.cpp)
-add_executable(${PROJECT_NAME}_Controller_Runner src/controller/Controller_Runner.cpp src/controller/Controller.cpp)
-add_executable(${PROJECT_NAME}_IMU_Broadcaster_Runner src/imu/IMU_Broadcaster_Runner.cpp src/imu/IMU_Broadcaster.cpp)
-add_executable(${PROJECT_NAME}_Transistor_LL_Broadcaster_Runner src/low_level/Transistor_LL_Broadcaster_Runner.cpp src/low_level/Transistor_LL_Broadcaster.cpp)
-add_executable(${PROJECT_NAME}_Transistor_GPS_Broadcaster_Runner src/gps/Transistor_GPS_Broadcaster_Runner.cpp src/gps/Transistor_GPS_Broadcaster.cpp)
-add_executable(${PROJECT_NAME}_Transistor_Localizer_Runner src/localizer/Localizer_Runner.cpp src/localizer/Localizer.cpp)
+add_executable(${PROJECT_NAME}_Transistor_Controller_Runner src/transistor/controller/Controller_Runner.cpp src/transistor/controller/Controller.cpp)
+add_executable(${PROJECT_NAME}_Transistor_IMU_Broadcaster_Runner src/transistor/imu/IMU_Broadcaster_Runner.cpp src/transistor/imu/IMU_Broadcaster.cpp)
+add_executable(${PROJECT_NAME}_Transistor_LL_Broadcaster_Runner src/transistor/low_level/LL_Broadcaster_Runner.cpp src/transistor/low_level/LL_Broadcaster.cpp)
+add_executable(${PROJECT_NAME}_Transistor_GPS_Broadcaster_Runner src/transistor/gps/GPS_Broadcaster_Runner.cpp src/transistor/gps/GPS_Broadcaster.cpp)
+add_executable(${PROJECT_NAME}_Transistor_Localizer_Runner src/transistor/localizer/Localizer_Runner.cpp src/transistor/localizer/Localizer.cpp)
 
-add_dependencies(${PROJECT_NAME}_Controller_Runner robobuggy_generate_messages_cpp)
-add_dependencies(${PROJECT_NAME}_IMU_Broadcaster_Runner robobuggy_generate_messages_cpp)
+add_dependencies(${PROJECT_NAME}_Transistor_Controller_Runner robobuggy_generate_messages_cpp)
+add_dependencies(${PROJECT_NAME}_Transistor_IMU_Broadcaster_Runner robobuggy_generate_messages_cpp)
 add_dependencies(${PROJECT_NAME}_Transistor_LL_Broadcaster_Runner robobuggy_generate_messages_cpp)
 add_dependencies(${PROJECT_NAME}_Transistor_GPS_Broadcaster_Runner robobuggy_generate_messages_cpp)
 add_dependencies(${PROJECT_NAME}_Transistor_Localizer_Runner robobuggy_generate_messages_cpp)
@@ -161,8 +161,8 @@ add_dependencies(${PROJECT_NAME}_Transistor_Localizer_Runner robobuggy_generate_
 ## target back to the shorter version for ease of user use
 ## e.g. "rosrun someones_pkg node" instead of "rosrun someones_pkg someones_pkg_node"
 # set_target_properties(${PROJECT_NAME}_node PROPERTIES OUTPUT_NAME node PREFIX "")
-set_target_properties(${PROJECT_NAME}_Controller_Runner PROPERTIES OUTPUT_NAME Controller PREFIX "")
-set_target_properties(${PROJECT_NAME}_IMU_Broadcaster_Runner PROPERTIES OUTPUT_NAME IMU_Broadcaster PREFIX "")
+set_target_properties(${PROJECT_NAME}_Transistor_Controller_Runner PROPERTIES OUTPUT_NAME Transistor_Controller PREFIX "")
+set_target_properties(${PROJECT_NAME}_Transistor_IMU_Broadcaster_Runner PROPERTIES OUTPUT_NAME Transistor_IMU_Broadcaster PREFIX "")
 set_target_properties(${PROJECT_NAME}_Transistor_LL_Broadcaster_Runner PROPERTIES OUTPUT_NAME Transistor_LL_Broadcaster PREFIX "")
 set_target_properties(${PROJECT_NAME}_Transistor_GPS_Broadcaster_Runner PROPERTIES OUTPUT_NAME Transistor_GPS_Broadcaster PREFIX "")
 set_target_properties(${PROJECT_NAME}_Transistor_Localizer_Runner PROPERTIES OUTPUT_NAME Transistor_Localizer PREFIX "")
@@ -172,13 +172,11 @@ set_target_properties(${PROJECT_NAME}_Transistor_Localizer_Runner PROPERTIES OUT
 # add_dependencies(${PROJECT_NAME}_node ${${PROJECT_NAME}_EXPORTED_TARGETS} ${catkin_EXPORTED_TARGETS})
 
 ## Specify libraries to link a library or executable target against
-target_link_libraries(${PROJECT_NAME}_Controller_Runner
+target_link_libraries(${PROJECT_NAME}_Transistor_Controller_Runner
   ${catkin_LIBRARIES}
 )
-target_link_libraries(${PROJECT_NAME}_IMU_Broadcaster_Runner
+target_link_libraries(${PROJECT_NAME}_Transistor_IMU_Broadcaster_Runner
   ${catkin_LIBRARIES}
-)
-target_link_libraries(${PROJECT_NAME}_IMU_Broadcaster_Runner
   freespace
 )
 target_link_libraries(${PROJECT_NAME}_Transistor_LL_Broadcaster_Runner
