+++ Software/real_time/ROS_RoboBuggy/src/robobuggy/include/transistor/gps/GPS_Broadcaster.h
@@ -1,33 +1,31 @@
 //
-// Created by bhai on 9/29/17.
+// Created by bhai on 9/15/17.
 //
 
-#ifndef ROS_ROBOBUGGY_TRANSISTOR_GPS_BROADCASTER_H
-#define ROS_ROBOBUGGY_TRANSISTOR_GPS_BROADCASTER_H
+#ifndef ROS_ROBOBUGGY_GPS_BROADCASTER_H
+#define ROS_ROBOBUGGY_GPS_BROADCASTER_H
 
 #include "ros/ros.h"
+#include "robobuggy/GPS.h"
 #include "serial/serial.h"
 
 #include <sstream>
 
-#include <robobuggy/GPS.h>
-
-class Transistor_GPS_Broadcaster
+class GPS_Broadcaster
 {
 public:
-    Transistor_GPS_Broadcaster();
+    GPS_Broadcaster();
     static const std::string NODE_NAME;
     int handle_serial_messages();
 private:
-    ros::Publisher gps_pub;
     ros::NodeHandle nh;
+    ros::Publisher gps_pub;
     std::string serial_port;
     int serial_baud;
     std::string gps_serial_buffer;
 
     double convert_to_latitude(std::string str);
-
     double convert_to_longitude(std::string str);
 };
 
-#endif //ROS_ROBOBUGGY_TRANSISTOR_GPS_BROADCASTER_H
+#endif //ROS_ROBOBUGGY_GPS_BROADCASTER_H
