+++ Software/real_time/ROS_RoboBuggy/src/robobuggy/launch/transistor.launch
@@ -4,7 +4,7 @@
 
     <!-- Node that translates messages from low-level into ROS messages -->
     <node name="Transistor_LL_Broadcaster" pkg="robobuggy" type="Transistor_LL_Broadcaster" required="true" output="screen">
-        <!-- Serial port on device for the Arduino Mega - TODO: udev rule? -->
+        <!-- Serial port on device for the Arduino Mega -->
         <param name="serial_port" type="string" value="/dev/buggyarduino"/>
         <!-- Baud rate for communicating with low level -->
         <param name="serial_baud" type="int" value="57600"/>
@@ -12,7 +12,7 @@
 
     <!-- Node that translates messages from the GPS into ROS messages -->
     <node name="Transistor_GPS_Broadcaster" pkg="robobuggy" type="Transistor_GPS_Broadcaster" required="true">
-        <!-- Serial port on device for the UBLOX gps- TODO: udev rule? -->
+        <!-- Serial port on device for the UBLOX gps -->
         <param name="serial_port" type="string" value="/dev/buggygps"/>
         <!-- Baud rate for communicating with low level -->
         <param name="serial_baud" type="int" value="9600"/>
